requirements.txt
streamlit
yfinance
pandas
numpy
alpaca-py
plotly
matplotlib
ta
redis
bcrypt
python-dotenv
psycopg2-binary
